requirements.txt
streamlit
docling
opensearch-py
google-genai
python-dotenv
pandas
plotly